metrics and evaluate script together
